requirements.txt
django==5.2.12
djangorestframework==3.16.1
markdown==3.10.2             # Markdown support for the browsable API.
django-cors-headers==4.9.0
